requirements.txt
requests
openai
python-dotenv
schedule
twython